requirements.txt
googletrans
Pillow
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.2.0+cu121
torchvision
torchaudio
transformers